% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_tool_immune}
\alias{add_tool_immune}
\alias{add_tool_mask}
\alias{add_tool_vaccine}
\alias{add_virus_covid19}
\alias{new_model}
\alias{run_model}
\title{Create a new model}
\usage{
add_tool_immune(model, baselineseq, preval)

add_tool_mask(model, baselineseq, preval)

add_tool_vaccine(model, baselineseq, preval)

add_virus_covid19(model, baselineseq, preval)

new_model()

run_model(model, nsteps, seed)
}
\description{
Create a new model
}
