% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_info}
\alias{get_info}
\alias{get_today_variant}
\alias{get_today_total}
\alias{get_hist_total}
\alias{get_hist_variant}
\title{Extract database information from the model
//'}
\usage{
get_today_variant(model)

get_today_total(model)

get_hist_total(model)

get_hist_variant(model)
}
\description{
Extract database information from the model
//'
}
