% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_epi_model}
\alias{new_epi_model}
\alias{init_epi_model}
\alias{run_epi_model}
\alias{print.epi_model}
\alias{edgelist_from_file}
\alias{edgelist_from_vec}
\alias{add_tool_immune}
\alias{add_tool_vaccine}
\alias{add_tool_mask}
\alias{add_virus_covid19}
\title{Creates a new model}
\usage{
new_epi_model()

init_epi_model(model, nsteps, seed)

run_epi_model(model)

\method{print}{epi_model}(x)

edgelist_from_file(model, fname, skip, directed)

edgelist_from_vec(model, source, target, directed)

add_tool_immune(model, baselineseq, preval)

add_tool_vaccine(model, baselineseq, preval)

add_tool_mask(model, baselineseq, preval)

add_virus_covid19(model, baselineseq, preval)
}
\description{
Creates a new model
}
